whole project in one day refactor challange

+++ MusicSignatureBuilder/Point.cs
@@ -2,11 +2,11 @@
 {
     public struct Point
     {
-        public double X { get; set; }
+        public float X { get; set; }
 
-        public double Y { get; set; }
+        public float Y { get; set; }
 
-        public Point(double x, double y)
+        public Point(float x, float y)
         {
             X = x;
             Y = y;
